feat: refactor - move editor
