--- products/editor/package.json
@@ -1,4 +1,0 @@
-{
-    "name": "@posthog/editor",
-    "peerDependencies": {}
-}
