+++ products/llm_observability/providers/formatters/openai_formatter.py
@@ -10,7 +10,7 @@ from openai.types.chat import (
     ChatCompletionMessageToolCallParam,
 )
 
-from products.editor.backend.providers.formatters.anthropic_typeguards import (
+from products.llm_observability.providers.formatters.anthropic_typeguards import (
     is_base64_image_param,
     is_image_block_param,
     is_text_block_param,
