+++ products/llm_observability/providers/inkeep.py
@@ -7,7 +7,7 @@ from anthropic.types import MessageParam
 import logging
 from typing import Any
 
-from products.editor.backend.providers.formatters.openai_formatter import convert_to_openai_messages
+from products.llm_observability.providers.formatters.openai_formatter import convert_to_openai_messages
 
 logger = logging.getLogger(__name__)
