+++ tach.toml
@@ -28,7 +28,6 @@ depends_on = [
 
     # NOTE: Add new product dependencies here and in settings/web.py PRODUCTS_APPS
     "products.early_access_features",
-    "products.editor",
     "products.links",
     "products.revenue_analytics",
     "products.user_interviews",
@@ -36,10 +35,6 @@ depends_on = [
 
 [[modules]]
 path = "products.early_access_features"
-depends_on = ["posthog"]
-
-[[modules]]
-path = "products.editor"
 depends_on = ["posthog"]
 
 [[modules]]
